requirements.txt
azureml-defaults
azureml-mlflow
azureml-core
azureml-sdk
azureml-widgets
azureml-dataprep
azureml-dataprep-native
azureml-dataprep-rslex
azureml-dataset-runtime
azureml-telemetry
azureml-pipeline
applicationinsights
imbalanced-learn==0.7.0
imblearn
lightgbm
pandas
papermill
seaborn
numpy
pillow==8.0.0